Handle semicolon-separated commands in PowerShell
PowerShell supports command lines like `rush install; rush build`, where the `;` separates commands. The `$cursorPosition` parameter reflects the index into the entire string, so that assumption that it is bounded by the length of the input is incorrect in PowerShell.

Bounding the value to the length of the input string gives a "good enough" approximation of the desired behavior.